refactor(nat): Further move stateless NAT code into subdirectory
Create a subdirectory for the stateless NAT implementation. Move the
iplist module into this directory, as it seems likely that only the
stateless mode will use it.

This is in case we need to move more code (from example, from the mgmt
crate) into the "stateless" module.
