+++ nat/src/lib.rs
@@ -40,9 +40,8 @@
 //! - The total number of available (not excluded) private addresses used in an "Expose" object must
 //!   be equal to the total number of publicly exposed addresses in this object.
 
-mod iplist;
 mod stateful;
-pub mod stateless;
+mod stateless;
 
 pub use stateful::StatefulNat;
 pub use stateless::StatelessNat;
